requirements.txt
langgraph==0.2.45
langchain==0.3.15
langchain-openai==0.2.14
langchain-anthropic==0.3.22
chromadb==0.5.23
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
fastapi==0.115.6
uvicorn==0.32.1
